requirements.txt
Flask==2.3.2
flask-restful==0.3.9
webargs==8.2.0
marshmallow==3.19.0
flask-apispec==0.11.4
elasticsearch-dsl==7.4.1
click==8.1.3
python-dotenv==1.0.0
pyyaml
rootpath==0.1.1
flake8==6.0.0
pytest==7.3.1
mock==5.0.2